refactor : extract Tmap script file
