+++ client/src/script/Tmap.tsx
@@ -1,0 +1,28 @@
+// @ts-ignore
+const { Tmapv3 } = window;
+
+const TmapObject = {
+
+    initTmap: () => {
+        return new Tmapv3.Map("TMapApp", {
+            center: new Tmapv3.LatLng(37.56520450, 126.98702028),
+            width: "300px",	// 지도의 넓이
+            height: "300px",	// 지도의 높이
+            zoom: 16	// 지도 줌레벨
+        });
+    },
+
+    setMarker: (map: any, lat: number, lon: number) => {
+        new Tmapv3.Marker({
+            position: new Tmapv3.LatLng(lat, lon),
+            map: map
+        });
+    },
+
+    setViewMap: (map: any, lat: number, lon: number) => {
+        map.setCenter(new Tmapv3.LatLng(lat, lon));
+        map.setZoom(16);
+    }
+}
+
+export default TmapObject;
